Make AqtScheduleConfig as optional when creating a TensorQuantizer.
Currently, users are required to pass AqtScheduleConfig to create a TensorQuantizer no matter if they actually want to quantize weights or activations. It would be bothering in case where only weights are quantized (now AqtScheduleConfig needs to be passed along with FloatConfig in it). This CL removes such a strict requirement and makes the config as optional when creating TensorQuantizers.

This change can also speed up training since there won't be any statistics computations for weights in case of weight only quantization.

PiperOrigin-RevId: 477284487